Removed code duplicated between Azure and AWS
